requirements.txt
pyyaml
numpy
scikit-learn
tqdm
openai
anthropic
textblob